% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ds.debug.get}
\alias{ds.debug.get}
\title{Retrieve Object from DataSHIELD Server}
\usage{
ds.debug.get(object, conns = NULL)
}
\arguments{
\item{object}{The name of the object to retrieve for debugging.}
}
\value{
The requested object from all connected servers, or error messages if encountered.
}
\description{
This function requests an object from the DataSHIELD server for debugging purposes.
}
